feat: about me load unload working: imgui about level instead of rendering sprites.

+++ SDLGame/managers/config.h
@@ -13,7 +13,7 @@ public:
 	
 	static void UpdateMaxScore( int highScore );
 
-	static glm::vec2 GetWindowSize() { return m_WindowSize; }
+	static glm::ivec2 GetWindowSize() { return m_WindowSize; }
 	static int GetWindowPadding() { return m_WindowPadding; };
 	static int GetHighScore() { return m_HighScore; };
